deployment/requirements.txt
fastapi
uvicorn
pillow
torch
torchvision
numpy
pandas
scikit-learn
# Add any other dependencies your src/ code needs
torch>=2.0.0
torchvision>=0.15.0
scikit-learn>=1.2.0
pandas>=2.0.0
numpy>=1.24.0
Pillow>=9.0.0
fastapi>=0.100.0
uvicorn>=0.22.0
gradio>=4.0.0
opencv-python>=4.7.0
matplotlib>=3.7.0
scipy>=1.10.0
python-multipart>=0.0.6
